Importer: widen stage_versions semver columns to bigint (#30)
Migration 0001 widened versions.semver_* to bigint, but the temp staging
table the importer COPYs into still declared integer. The first
nixpkgs-unstable commit with a date-stamped component (8b7dc2ca,
0.1.20260720092025) failed every import with "value out of range for
type integer" and blocked the #20 catch-up at seq 2757.

Adds a PGlite test that drives the real staging DDL with such a version.

+++ packages/indexer/src/importSql.ts
@@ -89,9 +89,11 @@ export const STAGE_VERSIONS_DDL = `
     version text NOT NULL,
     sort_key bytea NOT NULL,
     prerelease boolean NOT NULL,
-    semver_major integer,
-    semver_minor integer,
-    semver_patch integer,
+    -- bigint like versions (migration 0001): nixpkgs has date-stamped
+    -- components such as 0.1.20260720092025 that overflow int4.
+    semver_major bigint,
+    semver_minor bigint,
+    semver_patch bigint,
     semver_pre text
   ) ON COMMIT DROP
 `;
